# verlet list rebuild optimizations