--- a/src/containers/VerletListHelpers.h
+++ b/src/containers/VerletListHelpers.h
@@ -51,17 +51,16 @@ class VerletListHelpers {
     virtual void SoAFunctor(SoA &soa, bool /*newton3*/ = true) override {
       if (soa.getNumParticles() == 0) return;
 
-      Particle **const __restrict__ idptr = reinterpret_cast<Particle * *const>(soa.begin(Particle::AttributeNames::id));
+      Particle **const __restrict__ idptr = reinterpret_cast<Particle **const>(soa.begin(Particle::AttributeNames::id));
       double *const __restrict__ xptr = soa.begin(Particle::AttributeNames::posX);
       double *const __restrict__ yptr = soa.begin(Particle::AttributeNames::posY);
       double *const __restrict__ zptr = soa.begin(Particle::AttributeNames::posZ);
 
-      for (unsigned int i = 0; i < soa.getNumParticles(); ++i) {
+      size_t numPart = soa.getNumParticles();
+      for (unsigned int i = 0; i < numPart; ++i) {
         auto &currentList = _verletListsAoS.at(idptr[i]);
 
-        for (unsigned int j = i + 1; j < soa.getNumParticles(); ++j) {
-          if (i == j) continue;
-
+        for (unsigned int j = i + 1; j < numPart; ++j) {
           const double drx = xptr[i] - xptr[j];
           const double dry = yptr[i] - yptr[j];
           const double drz = zptr[i] - zptr[j];
@@ -82,20 +81,25 @@ class VerletListHelpers {
     virtual void SoAFunctor(SoA &soa1, SoA &soa2, bool /*newton3*/ = true) override {
       if (soa1.getNumParticles() == 0 || soa2.getNumParticles() == 0) return;
 
-      Particle **const __restrict__ id1ptr = reinterpret_cast<Particle * *const>(soa1.begin(Particle::AttributeNames::id));
+      Particle **const __restrict__ id1ptr =
+          reinterpret_cast<Particle **const>(soa1.begin(Particle::AttributeNames::id));
       double *const __restrict__ x1ptr = soa1.begin(Particle::AttributeNames::posX);
       double *const __restrict__ y1ptr = soa1.begin(Particle::AttributeNames::posY);
       double *const __restrict__ z1ptr = soa1.begin(Particle::AttributeNames::posZ);
 
-      Particle **const __restrict__ id2ptr = reinterpret_cast<Particle * *const>(soa2.begin(Particle::AttributeNames::id));
+      Particle **const __restrict__ id2ptr =
+          reinterpret_cast<Particle **const>(soa2.begin(Particle::AttributeNames::id));
       double *const __restrict__ x2ptr = soa2.begin(Particle::AttributeNames::posX);
       double *const __restrict__ y2ptr = soa2.begin(Particle::AttributeNames::posY);
       double *const __restrict__ z2ptr = soa2.begin(Particle::AttributeNames::posZ);
 
-      for (unsigned int i = 0; i < soa1.getNumParticles(); ++i) {
+      size_t numPart1 = soa1.getNumParticles();
+      for (unsigned int i = 0; i < numPart1; ++i) {
         auto &currentList = _verletListsAoS.at(id1ptr[i]);
 
-        for (unsigned int j = 0; j < soa2.getNumParticles(); ++j) {
+        size_t numPart2 = soa2.getNumParticles();
+
+        for (unsigned int j = 0; j < numPart2; ++j) {
           const double drx = x1ptr[i] - x2ptr[j];
           const double dry = y1ptr[i] - y2ptr[j];
           const double drz = z1ptr[i] - z2ptr[j];
@@ -127,8 +131,8 @@ class VerletListHelpers {
       auto cellIter = cell.begin();
       // load particles in SoAs
       for (size_t i = 0; cellIter.isValid(); ++cellIter, ++i) {
-        Particle* pptr = &(*cellIter);
-        idptr[i] = reinterpret_cast<double&>(pptr);
+        Particle *pptr = &(*cellIter);
+        idptr[i] = reinterpret_cast<double &>(pptr);
         xptr[i] = cellIter->getR()[0];
         yptr[i] = cellIter->getR()[1];
         zptr[i] = cellIter->getR()[2];