the verlet-lists are now (re)build using SoA's.
This saves a lot of lookups (hash generations) and some other things.

Speedup is clearly visible, especially as soon as there are more than one particle per cell:
Verlet Lists (md-main, 1000 cells):

type | particle count | rebuild time before | rebuild time now | kernel time (LJ)
------ | ----------------- | ------------------------ | ---------------------- | -----------------
soa | 10.000 | 0.12 | 0.07 | 0.010
aos | 10.000 | 0.10 | 0.05 | 0.018
soa | 1.000 | 1.8e-3 | 1.4e-3 | 1.9e-4
aos | 1.000 | 1.6e-3 | 1.2e-3 | 1.8e-4

Linked Cells (md-main, 1000 cells):

type | particle count | kernel time (LJ)
------ | --------------- | ----
soa | 10.000 | 0.032
aos | 10.000 | 0.07
soa | 1.000 | 6.8e-4
aos | 1.000 | 1.0e-3

meaning, that in this case verlet lists are faster than linked cells, as soon as the rebuild happens at most every 4 time steps.

remaining optimization possibilities:
* from what I see, vectorization is not really possible here, as there are not really any flops done. the main time is spent in storing the lists
* upper optimization limit for the rebuild time should be roughly that of the kernel time of the aos LJ  linked-cells iteration. so we are already close to optimal.
* put neighborlists into particle class to prevent lookup alltogether -- involves lots of template magic and might not pay off