@SteffenSeckler 2018-06-05
@FG-TUM review here, too :)